Revert "Split parser package into topdown and bottomup (#250)"
This reverts commit 82877c523d9c09d00df82944fad5e1607bde07b8.
