+++ parser/predictive/example_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/moorara/algo/grammar"
 	"github.com/moorara/algo/lexer"
 	"github.com/moorara/algo/lexer/input"
-	"github.com/moorara/algo/parser/topdown/predictive"
+	"github.com/moorara/algo/parser/predictive"
 )
 
 type exprLexer struct {
